+++ regimes/no/no.go
@@ -1,0 +1,72 @@
+// Package no provides the tax regime definition for Norway.
+package no
+
+import (
+	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/currency"
+	"github.com/invopop/gobl/i18n"
+	"github.com/invopop/gobl/l10n"
+	"github.com/invopop/gobl/org"
+	"github.com/invopop/gobl/tax"
+)
+
+func init() {
+	tax.RegisterRegimeDef(New())
+}
+
+// New instantiates a new Norwegian tax regime.
+func New() *tax.RegimeDef {
+	return &tax.RegimeDef{
+		Country:   l10n.NO.Tax(),
+		Currency:  currency.NOK,
+		TaxScheme: tax.CategoryVAT,
+		Name: i18n.String{
+			i18n.EN: "Norway",
+			i18n.NB: "Norge",
+		},
+		TimeZone:   "Europe/Oslo",
+		Identities: identityTypeDefinitions,
+		Categories: taxCategories,
+		Tags: []*tax.TagSet{
+			invoiceTags,
+		},
+		Corrections: []*tax.CorrectionDefinition{
+			{
+				Schema: bill.ShortSchemaInvoice,
+				Types: []cbc.Key{
+					bill.InvoiceTypeCreditNote,
+					bill.InvoiceTypeDebitNote,
+				},
+			},
+		},
+		Scenarios: []*tax.ScenarioSet{
+			invoiceScenarios(),
+		},
+		Validator:  Validate,
+		Normalizer: Normalize,
+	}
+}
+
+// Validate checks the document type and determines if it can be validated.
+func Validate(doc any) error {
+	switch obj := doc.(type) {
+	case *bill.Invoice:
+		return validateBillInvoice(obj)
+	case *tax.Identity:
+		return validateTaxIdentity(obj)
+	case *org.Identity:
+		return validateOrgIdentity(obj)
+	}
+	return nil
+}
+
+// Normalize performs any regime-specific normalizations.
+func Normalize(doc any) {
+	switch obj := doc.(type) {
+	case *tax.Identity:
+		normalizeTaxIdentity(obj)
+	case *org.Identity:
+		normalizeOrgIdentity(obj)
+	}
+}